+++ src/main/java/com/atlauncher/data/mojang/MojangDownload.java
@@ -15,22 +15,25 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
+package com.atlauncher.data.mojang;
 
-package com.atlauncher;
+import com.atlauncher.annot.Json;
 
-public final class ExceptionStrainer implements Thread.UncaughtExceptionHandler {
-    @Override
-    public void uncaughtException(Thread t, Throwable e) {
-        e.printStackTrace();
+@Json
+public class MojangDownload {
+    private String sha1;
+    private long size;
+    private String url;
 
-        if (e.getMessage() != null && !e.getMessage().isEmpty()) {
-            LogManager.error(e.getMessage());
-        }
+    public String getSha1() {
+        return this.sha1;
+    }
 
-        for (StackTraceElement element : e.getStackTrace()) {
-            if (element != null) {
-                LogManager.error(element.toString());
-            }
-        }
+    public long getSize() {
+        return this.size;
+    }
+
+    public String getUrl() {
+        return this.url;
     }
 }