+++ src/main/java/com/atlauncher/data/mojang/MojangDownloads.java
@@ -15,22 +15,25 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
+package com.atlauncher.data.mojang;
 
-package com.atlauncher;
+import com.atlauncher.annot.Json;
 
-public final class ExceptionStrainer implements Thread.UncaughtExceptionHandler {
-    @Override
-    public void uncaughtException(Thread t, Throwable e) {
-        e.printStackTrace();
+@Json
+public class MojangDownloads {
+    private MojangDownload client;
+    private MojangDownload server;
+    private MojangDownload windows_server;
 
-        if (e.getMessage() != null && !e.getMessage().isEmpty()) {
-            LogManager.error(e.getMessage());
-        }
+    public MojangDownload getClient() {
+        return this.client;
+    }
 
-        for (StackTraceElement element : e.getStackTrace()) {
-            if (element != null) {
-                LogManager.error(element.toString());
-            }
-        }
+    public MojangDownload getServer() {
+        return this.server;
+    }
+
+    public MojangDownload getWindowsServer() {
+        return this.windows_server;
     }
 }