+++ src/main/java/com/atlauncher/gui/dialogs/ModsChooser.java
@@ -54,6 +54,7 @@ public class ModsChooser extends JDialog {
     private JButton clearAllButton;
     private JButton installButton;
     private List<ModsJCheckBox> modCheckboxes;
+    private List<ModsJCheckBox> sortedOut;
 
     private boolean wasClosed = false;
 
@@ -325,7 +326,7 @@ public class ModsChooser extends JDialog {
             checkBox.addActionListener(new ActionListener() {
                 public void actionPerformed(ActionEvent e) {
                     ModsJCheckBox a = (ModsJCheckBox) e.getSource();
-                    sortOutMods(a);
+                    sortOutMods(a, true);
                 }
             });
             modCheckboxes.add(checkBox);
@@ -341,6 +342,14 @@ public class ModsChooser extends JDialog {
                 checkBoxPanel2.add(checkBox);
             }
         }
+
+        sortedOut = new ArrayList<ModsJCheckBox>();
+        for (ModsJCheckBox cb : this.modCheckboxes) {
+            if ((installer.isServer() ? cb.getMod().isServerOptional() : cb.getMod().isOptional()) && cb.isSelected()) {
+                sortOutMods(cb);
+            }
+        }
+
         checkBoxPanel1.setPreferredSize(new Dimension(0, count1 * 20));
         checkBoxPanel2.setPreferredSize(new Dimension(0, count2 * 20));
 
@@ -435,6 +444,14 @@ public class ModsChooser extends JDialog {
     }
 
     public void sortOutMods(ModsJCheckBox a) {
+        this.sortOutMods(a, false);
+    }
+
+    public void sortOutMods(ModsJCheckBox a, boolean firstGo) {
+        if (firstGo) {
+            sortedOut = new ArrayList<ModsJCheckBox>();
+        }
+
         if (a.isSelected()) {
             List<Mod> linkedMods = modsToChange(a.getMod());
             for (Mod mod : linkedMods) {
@@ -458,8 +475,10 @@ public class ModsChooser extends JDialog {
                 List<Mod> dependsMods = modsDependancies(a.getMod());
                 for (Mod mod : dependsMods) {
                     for (ModsJCheckBox check : modCheckboxes) {
-                        if (check.getMod() == mod) {
+                        if (check.getMod() == mod && !sortedOut.contains(check)) {
+                            sortedOut.add(check);
                             check.setSelected(true);
+                            sortOutMods(check);
                         }
                     }
                 }