+++ src/main/java/com/atlauncher/utils/Utils.java
@@ -17,23 +17,6 @@
  */
 package com.atlauncher.utils;
 
-import com.atlauncher.App;
-import com.atlauncher.Gsons;
-import com.atlauncher.LogManager;
-import com.atlauncher.data.Constants;
-import com.atlauncher.data.mojang.ExtractRule;
-import com.atlauncher.data.mojang.OperatingSystem;
-import com.atlauncher.data.openmods.OpenEyeReportResponse;
-import com.atlauncher.evnt.LogEvent.LogType;
-import org.tukaani.xz.XZInputStream;
-
-import javax.crypto.BadPaddingException;
-import javax.crypto.Cipher;
-import javax.crypto.IllegalBlockSizeException;
-import javax.crypto.spec.SecretKeySpec;
-import javax.imageio.ImageIO;
-import javax.net.ssl.HttpsURLConnection;
-import javax.swing.ImageIcon;
 import java.awt.Desktop;
 import java.awt.Dimension;
 import java.awt.Font;
@@ -96,6 +79,25 @@ import java.util.regex.Pattern;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 import java.util.zip.ZipOutputStream;
+
+import javax.crypto.BadPaddingException;
+import javax.crypto.Cipher;
+import javax.crypto.IllegalBlockSizeException;
+import javax.crypto.spec.SecretKeySpec;
+import javax.imageio.ImageIO;
+import javax.net.ssl.HttpsURLConnection;
+import javax.swing.ImageIcon;
+
+import org.tukaani.xz.XZInputStream;
+
+import com.atlauncher.App;
+import com.atlauncher.Gsons;
+import com.atlauncher.LogManager;
+import com.atlauncher.data.Constants;
+import com.atlauncher.data.mojang.ExtractRule;
+import com.atlauncher.data.mojang.OperatingSystem;
+import com.atlauncher.data.openmods.OpenEyeReportResponse;
+import com.atlauncher.evnt.LogEvent.LogType;
 
 public class Utils {
     public static String error(Throwable t) {
@@ -164,10 +166,14 @@ public class Utils {
     }
 
     public static File getCoreGracefully() {
+        if (App.workingDir != null) {
+            return App.workingDir;
+        }
+
         if (Utils.isLinux()) {
             try {
                 return new File(App.class.getProtectionDomain().getCodeSource().getLocation().toURI()
-                        .getSchemeSpecificPart()).getParentFile();
+                    .getSchemeSpecificPart()).getParentFile();
             } catch (URISyntaxException e) {
                 e.printStackTrace();
                 return new File(System.getProperty("user.dir"), Constants.LAUNCHER_NAME);
@@ -395,6 +401,15 @@ public class Utils {
      */
     public static boolean is64Bit() {
         return System.getProperty("sun.arch.data.model").contains("64");
+    }
+
+    /**
+     * Checks if Windows is 64 bit
+     *
+     * @return true, if it is 64 bit
+     */
+    public static boolean isWindows64Bit() {
+        return System.getenv("ProgramFiles(x86)") != null;
     }
 
     /**
@@ -711,11 +726,11 @@ public class Utils {
     public static boolean copyFile(File from, File to, boolean withFilename) {
         if (!from.isFile()) {
             LogManager.error("File " + from.getAbsolutePath() + " cannot be copied to " + to.getAbsolutePath() + " as" +
-                    " it isn't a file");
+                " it isn't a file");
         }
         if (!from.exists()) {
             LogManager.error("File " + from.getAbsolutePath() + " cannot be copied to " + to.getAbsolutePath() + " as" +
-                    " it doesn't exist");
+                " it doesn't exist");
             return false;
         }
         if (!withFilename) {
@@ -801,7 +816,7 @@ public class Utils {
             return true;
         } else {
             LogManager.error("Couldn't move directory " + sourceLocation.getAbsolutePath() + " to " + targetLocation
-                    .getAbsolutePath());
+                .getAbsolutePath());
             return false;
         }
     }
@@ -952,13 +967,13 @@ public class Utils {
 
         if (isSymlink(file)) {
             LogManager.error("Not deleting the " + (file.isFile() ? "file" : "folder") + file.getAbsolutePath() +
-                    "as it's a symlink");
+                "as it's a symlink");
             return;
         }
 
         if (!file.delete()) {
             LogManager.error((file.isFile() ? "File" : "Folder") + " " + file.getAbsolutePath() + " couldn't be " +
-                    "deleted");
+                "deleted");
         }
     }
 
@@ -1018,7 +1033,7 @@ public class Utils {
             } else {
                 String hash = getSHA1(file);
                 File saveTo = new File(App.settings.getObjectsAssetsDir(), hash.substring(0, 2) + File.separator +
-                        hash);
+                    hash);
                 saveTo.mkdirs();
                 copyFile(file, saveTo, true);
             }
@@ -1213,7 +1228,7 @@ public class Utils {
      * @throws IOException Signals that an I/O exception has occurred.
      */
     public static void replaceText(File originalFile, File destinationFile, String replaceThis, String withThis)
-            throws IOException {
+        throws IOException {
 
         FileInputStream fs = new FileInputStream(originalFile);
         BufferedReader br = new BufferedReader(new InputStreamReader(fs));
@@ -1459,7 +1474,7 @@ public class Utils {
             return "Launcher: " + System.getProperty("java.version") + ", Minecraft: " + version;
         } else {
             return "Launcher: " + System.getProperty("java.version") + ", Minecraft: " + System.getProperty("java" +
-                    ".version");
+                ".version");
         }
     }
 
@@ -1624,7 +1639,7 @@ public class Utils {
         String request = Utils.getFileContents(report);
         if (request == null) {
             LogManager.error("OpenEye: Couldn't read contents of file '" + report.getAbsolutePath() + "'. Pending " +
-                    "report sending failed!");
+                "report sending failed!");
             return null;
         }
 
@@ -2051,7 +2066,7 @@ public class Utils {
 
         int x = decompressed.length;
         int len = ((decompressed[x - 8] & 0xFF)) | ((decompressed[x - 7] & 0xFF) << 8) | ((decompressed[x - 6] &
-                0xFF) << 16) | ((decompressed[x - 5] & 0xFF) << 24);
+            0xFF) << 16) | ((decompressed[x - 5] & 0xFF) << 24);
         byte[] checksums = Arrays.copyOfRange(decompressed, decompressed.length - len - 8, decompressed.length - 8);
         try {
             FileOutputStream jarBytes = new FileOutputStream(output);